<dce.h> : must be "unsigned int" instead of "unsigned long (int)"

+++ freedce/include/dce/linux-gnu/dce.h
@@ -164,25 +164,25 @@ extern const char *dceshared_path;
 typedef idl_byte        byte;
 typedef idl_boolean     boolean;
 typedef struct unsigned64_s_t {
-    unsigned long hi;
-    unsigned long lo;
+    unsigned int hi;
+    unsigned int lo;
 } unsigned64;
 
 typedef struct signed64_s_t {
-    unsigned long hi;
-    unsigned long lo;
+    unsigned int hi;
+    unsigned int lo;
 } signed64;
 
 typedef struct unsigned48_s_t {
-    unsigned long  int  lo;             /* least significant 32 bits */
-	unsigned short int  hi;             /* most significant 16 bits */
+    unsigned int   lo;             /* least significant 32 bits */
+	unsigned short hi;             /* most significant 16 bits */
 } unsigned48;
 
 typedef struct unsigned128_s_t {
-    unsigned long lolo;
-    unsigned long lohi;
-    unsigned long hilo;
-    unsigned long hihi;
+    unsigned int lolo;
+    unsigned int lohi;
+    unsigned int hilo;
+    unsigned int hihi;
 } unsigned128;
 
 /*
